% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cojam}
\alias{cojam}
\title{Merge two \code{\link[R2BGLiMS]{JAM}} models}
\usage{
cojam(jam_res_1, jam_res_2, prior_odds_43 = NA)
}
\arguments{
\item{jam_res_1}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object,
from running \code{\link[R2BGLiMS]{JAM}} for trait 1.}

\item{jam_res_2}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object,
from running \code{\link[R2BGLiMS]{JAM}} for trait 2.}

\item{prior_odds_43}{Optional vector of prior odds of a shared variant ("H4") versus
distinct variants ("H3"). Can also be added later using \code{\link{cojam_odds}}.}
}
\value{
A \code{list} containing the following elements:
  \describe{
      \item{\code{bayes_factor}}{Bayes Factor for H4 versus H3.}
      \item{\code{posterior_overlap}}{Posterior probability of H4 U H3.}
      \item{\code{posterior_odds_43}}{If \code{prior_odds_43} are provided,
      the posterior odds of H4 versus H3.}
      \item{\code{model_info}}{
      \describe{Extra info for other functions / post-processing:
        \item{\code{models_grid}}{All sampled configurations (joint
          \code{\link[R2BGLiMS]{JAM}} models) and their posterior counts.}
        \item{\code{implied_prior_odds_43}}{Prior odds of H4 versus H3 that would
          have been implied under an assumption of independence between traits.
          Provided as a sanity check for user-supplied priors, which should
          usually promote H4 vs H3 (i.e. user supplied \code{prior_odds_43}
          should usually be greater than \code{implied_prior_odds_43}).}
        }
      }
    }
}
\description{
Combines samples from the posterior model space of two \code{\link[R2BGLiMS]{JAM}}
models, allowing posterior inference about colocalisation, i.e. about the
existence of one or more shared variants that are associated with both outcome
traits.
}
