% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cojam_odds}
\alias{cojam_odds}
\title{Augment \code{\link{cojam}} results with other prior choices}
\usage{
cojam_odds(cojam_res, prior_odds_43)
}
\arguments{
\item{cojam_res}{Results from \code{\link{cojam}}}

\item{prior_odds_43}{Vector of prior odds of H4 versus H3.}
}
\value{
Same as \code{\link{cojam}}, a \code{list} containing the following elements:
  \describe{
      \item{\code{bayes_factor}:}{Bayes Factor for H4 versus H3.}
      \item{\code{posterior_overlap}:}{Posterior probability of H4 U H3.}
      \item{\code{posterior_odds_43}:}{Posterior odds of H4 versus H3, including
      those already in \code{cojam_res} (if any) and new rows for new \code{prior_odds_43}.}
    }
}
\description{
Calculate posterior odds of a shared variant ("H4") versus distinct variants ("H3")
}
