% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{complement_DNA}
\alias{complement_DNA}
\title{Convert DNA string to its complement}
\usage{
complement_DNA(x)
}
\arguments{
\item{x}{Character string of nucleotide code}
}
\value{
Character string of the complement of \code{x}
}
\description{
Convert string of nucleotide codes to its complement
}
