% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_coloc_data}
\alias{convert_coloc_data}
\title{Convert \code{\link[coloc]{coloc.abf}} data to \code{\link[R2BGLiMS]{JAM}} format}
\usage{
convert_coloc_data(coloc_data, ref_genotypes)
}
\arguments{
\item{coloc_data}{Data formatted for \code{\link[coloc]{coloc.abf}}}

\item{ref_genotypes}{Reference genotype matrix used by JAM to impute the SNP-SNP
correlations. Genotypes must be coded as a numeric risk allele
count 0/1/2. Non-integer values reflecting imputation may be given. NB: The
risk allele coding must correspond to that used in \code{coloc_data}. Must be
positive definite, with SNP identifiers in the column names.}
}
\description{
Convert \code{\link[coloc]{coloc.abf}} data to the format required by
\code{\link[R2BGLiMS]{JAM}}, using \code{\link{jam_wrap}}.
}
