% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{implied_prior}
\alias{implied_prior}
\title{Calculate implied priors for two independent \code{\link[R2BGLiMS]{JAM}} models}
\usage{
implied_prior(jam_res_1, jam_res_2)
}
\arguments{
\item{jam_res_1}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object,
from running \code{\link[R2BGLiMS]{JAM}} for trait 1.}

\item{jam_res_2}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object,
from running \code{\link[R2BGLiMS]{JAM}} for trait 2.}
}
\value{
Named vector of implied prior probabilities for H0 U H1 U H2, H3 and H4.
}
\description{
For internal use by \code{\link{cojam}}. Uses combinatorics and each
  \code{\link[R2BGLiMS]{JAM}} model's beta-binomial prior to calculate the prior
  probabilities of each colocalisation hypothesis, given that the two traits
  are independent. Used by \code{\link{cojam}} to calculate the Bayes
  Factor for for H4 versus H3 (by factoring out this unrealistic implied prior).
}
