% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{jam_models}
\alias{jam_models}
\title{Tabulate all models visited by \code{\link[R2BGLiMS]{JAM}}}
\usage{
jam_models(jam_res)
}
\arguments{
\item{jam_res}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object,
from running \code{\link[R2BGLiMS]{JAM}}.}
}
\value{
A \code{\link[tibble]{tibble}} with one row for each visited model,
  with indicators showing inclusion of each SNP in that model, a count of
  posterior samples of that model (i.e. how often the rjMCMC visited it),
  and a ranking of the visited models.
}
\description{
For internal use by \code{\link{cojam}}.
}
