% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{jam_wrap}
\alias{jam_wrap}
\title{Wrapper around \code{\link[R2BGLiMS]{JAM}}}
\usage{
jam_wrap(
  marginal_beta,
  snp_names,
  ref_genotypes,
  n,
  trait_variance,
  binary_outcome = FALSE,
  marginal_beta_se = NULL,
  bb_prior_a = 1,
  bb_prior_b = NULL,
  ...
)
}
\arguments{
\item{marginal_beta}{Vector of marginal effect estimates to re-analyse with
JAM under multivariate models. For GWAS summaries, these or _log-ORs_.}

\item{snp_names}{SNP identifiers (e.g. RSID), in the same order as \code{marginal_beta}.}

\item{ref_genotypes}{Reference genotype matrix used by JAM to impute the SNP-SNP
correlations. Individual's genotype must be coded as a numeric risk allele
count 0/1/2. Non-integer values reflecting imputation may be given. NB: The
risk allele coding MUST correspond to that used in marginal.betas. Must be
positive definite, with SNP identifiers in the column names.}

\item{n}{The size of the dataset in which the summary statistics
\code{marginal.betas} were calculated}

\item{trait_variance}{Estimate of the trait (outcome) variance.}

\item{binary_outcome}{Is the trait (outcome) binary? Should be \code{TRUE} for
GWAS or \code{FALSE} for QTL studies.}

\item{marginal_beta_se}{Only required if the trait (outcome) is binary:
standard errors of the log-ORs.}

\item{bb_prior_a}{Parameter of Beta(a, b) prior on proportion of causal
covariates, default 1.}

\item{bb_prior_b}{Parameter of Beta(a, b) prior on proportion of causal
covariates. Default is the number of SNPs that are common to both \code{snp_names}
and \code{colnames(ref_genotypes)}. Higher values of \code{bb_prior_b}
relative to \code{bb_prior_a} will encourage greater sparsity. Use
\code{\link[R2BGLiMS]{GetBetaBinomParams}} for suggested beta-binomial parameters.}

\item{...}{Other arguments to \code{\link[R2BGLiMS]{JAM}}.}
}
\value{
A list of arguments that can be passed to \code{\link[R2BGLiMS]{JAM}}
  via \code{do.call}, or as an argument to \code{\link{cojam}}.
}
\description{
Constructs a list of arguments that can be passed to \code{\link[R2BGLiMS]{JAM}}
  via \code{do.call}, or as an argument to \code{\link{cojam}}. The prior
  proportion of causal covariates is treated as unknown and given a Beta(a, b)
  hyper-prior.
}
