% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_size_priors}
\alias{model_size_priors}
\title{Beta-Binomial probabilities for specific models}
\usage{
model_size_priors(jam_args)
}
\value{
Vector of the same length P (number of SNPs in \code{jam_args}),
  containing prior probabilities for models of dimensions 0 to (P-1).
}
\description{
This is not the prior on a particular dimension (that would required the
  binomial co-efficient). For use in calculating the implied prior (under
  independence between traits) in \code{cojam}.
}
