% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Plot marginal SNP posterior probability}
\usage{
plot_manhattan(cojam_res)
}
\arguments{
\item{cojam_res}{Results from \code{\link{cojam}}}
}
\description{
Three Manhattan plots: one for each trait (showing the marginal SNP posterior
  probabilities), and one for the marginal posterior probabilities of each
  SNP being shared for both traits.
}
