% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prune_tags}
\alias{prune_tags}
\title{Hierarchical pruning of genotype matrix}
\usage{
prune_tags(genotypes_1, genotypes_2 = NULL, threshold = 0.9)
}
\arguments{
\item{genotypes_1}{Genotype matrix, with SNP identifiers as column names.}

\item{genotypes_2}{Optional second genotype matrix, useful to jointly select
tag SNPs for use in \code{cojam}.}

\item{threshold}{Correlation (height) at which to cut the hierarchical
clustering tree.}
}
\value{
Character vector of tag SNP identifiers.
}
\description{
Find SNPs that "tag" groups of correlated SNPs by cutting a hierarchical
  clustering tree at the specified threshold.
}
