% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cojam}
\alias{cojam}
\title{Merge two \code{\link{R2BGLiS::JAM}} models}
\usage{
cojam(jam_args_1, jam_args_2, prior_odds_43 = NA)
}
\arguments{
\item{jam_args_1}{A \code{list} of arguments to JAM for trait 1. This can be
conveniently constructed using \code{\link{jam_wrap}}. See \code{\link{jam_wrap}}
and \code{\link[R2BGLiMS]{JAM}} for details of the required arguments.}

\item{jam_args_2}{A \code{list} of arguments to JAM for trait 2.}

\item{prior_odds_43}{Optional (vector of) prior odds of a shared variant ("H4") versus
distinct variants ("H3"). Can also be added later using \code{\link{cojam_odds}}.}
}
\value{
A \code{list} containing the following elements:
  \describe{
      \item{\code{bayes_factor}:}{Bayes Factor for H4 versus H3.}
      \item{\code{posterior_overlap}:}{Posterior probability of H4 U H3.}
      \item{\code{posterior_odds_43}:}{If \code{prior_odds_43} are provided,
      the posterior odds of H4 versus H3.}
    }
}
\description{
Combines samples from the posterior model space of two \code{\link[R2BGLiMS]{JAM}}
models, allowing posterior inference about colocalisation, i.e. about the
existence of one or more shared variants that are associated with both outcome
traits.
}
