% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_size_priors}
\alias{model_size_priors}
\title{Beta-Binomial probabilities for specific models}
\usage{
model_size_priors(jam_res)
}
\arguments{
\item{jam_res}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object,
from running \code{\link[R2BGLiMS]{JAM}}.}
}
\value{
Vector of the same length P (number of SNPs in \code{jam_res}),
  containing prior probabilities for models of dimensions 0 to (P-1).
}
\description{
For internal use by \code{\link{cojam}}.This is not the prior on a particular
  dimension (that would required the binomial co-efficient). For use in
  calculating the implied prior (under independence between traits) in \code{cojam}.
}
