% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{posterior_summaries}
\alias{posterior_summaries}
\title{Posterior summaries from \code{\link{cojam}} results}
\usage{
posterior_summaries(cojam_res, prior_odds_43)
}
\arguments{
\item{cojam_res}{Results from \code{\link{cojam}}}

\item{prior_odds_43}{Vector of prior odds of H4 versus H3.}
}
\value{
A \code{\link[tibble]{tibble}} of prior and posterior odds of H4
  versus H3, and posterior probabilities of H3 and H4.
}
\description{
Calculate posterior odds of a shared variant ("H4") versus distinct variants
("H3"), and posterior probability of H4, given user-defined prior odds of H4
versus H3.
}
